+++ doc/whats-new.rst
@@ -30,6 +30,8 @@ Bug fixes
 - Improved error handling and documentation for `.expand_dims()` 
   read-only view.
 
+- Add a DataArray.broadcast_like() method.
+
 .. _whats-new.0.12.3:
 
 v0.12.3 (10 July 2019)